Config.yml issues #4
Change the location of Config.yml to the folder probe
Change the tag sensor to the tag actions.
We can assume than we will use only minutes to mesure the periodicity.
For example: 15, 60 or 1440
I should remove the tags called type and version. We can create those
tags when we will need them.